Complete GameBoard class and gameplay logic
Oops forgot to commit this as it was being developed tonight.
Started with implementation of GameBoard class which initializes
a game board and prints it to the CLI. Then wrote a play_turn method
that takes input from the user, ensures its valid, and maps that input
to a cell on the game board. Need to break out valid input checking as
a helper method. Lastly implemented a check_for_winner method that
is called after each turn and checks all possible winning conditions
to see if either player has won the game yet.

Need to verify that all cases work properly. Also need to end the game
in a draw if the board fills up with no winner. Would be nice to fix
the formatting of occupied cells too.