Complete codebreaker game mode
Added String class extension methods and print_with_color method to be
called on color strings displayed to the UI. In combination, these
methods display strings with the corresponding color as a block
background to improve readability.

Added give_player_feedback method that evaluates the players guess
and returns the appropriate feedback indicating how accurate the
players guess was.

Developed game menu.

Added game play logic that manages menu selection and allows a game to
be played. Game begins with a random code being generated by the
computer and then 12 rounds of user guesses followed up with computer
feedback. The game ends when the user guesses the code correctly or runs
out of guesses.