Breakout codemaker game mode as method
First step in cleaning up the game code. Removed codemaker game mode
logic from main body of game.rb and broke it out as a separate method
to be called when codemaker game mode is selected by user at the main
menu. Likely to be moved to a separate file to improve cleanliness.