Improve AI of computer in codemaker mode
Made improvements to the computer's memory of correct colors guessed
which significantly improved likelihood of computer to guess the correct
code within 12 tries. The new algorithm, in addition to remembering
colors guessed in the correct position, remembers colors guessed in the
wrong position and includes those colors in subsequent guesses.